+++ chatlas/__init__.py
@@ -4,16 +4,22 @@ from ._auto import ChatAuto
 from ._chat import Chat
 from ._content_image import content_image_file, content_image_plot, content_image_url
 from ._github import ChatGithub
-from ._google import ChatGoogle
+from ._google import ChatGoogle, ChatVertex
 from ._groq import ChatGroq
 from ._interpolate import interpolate, interpolate_file
 from ._ollama import ChatOllama
 from ._openai import ChatAzureOpenAI, ChatOpenAI
 from ._perplexity import ChatPerplexity
 from ._provider import Provider
+from ._snowflake import ChatSnowflake
 from ._tokens import token_usage
 from ._tools import Tool
 from ._turn import Turn
+
+try:
+    from ._version import version as __version__
+except ImportError:  # pragma: no cover
+    __version__ = "0.0.0"  # stub value for docs
 
 __all__ = (
     "ChatAnthropic",
@@ -26,6 +32,8 @@ __all__ = (
     "ChatOpenAI",
     "ChatAzureOpenAI",
     "ChatPerplexity",
+    "ChatSnowflake",
+    "ChatVertex",
     "Chat",
     "content_image_file",
     "content_image_plot",