+++ chatlas/types/openai/_submit.py
@@ -38,12 +38,16 @@ class SubmitInputArgs(TypedDict, total=False):
     model: Union[
         str,
         Literal[
+            "o3-mini",
+            "o3-mini-2025-01-31",
             "o1",
             "o1-2024-12-17",
             "o1-preview",
             "o1-preview-2024-09-12",
             "o1-mini",
             "o1-mini-2024-09-12",
+            "gpt-4.5-preview",
+            "gpt-4.5-preview-2025-02-27",
             "gpt-4o",
             "gpt-4o-2024-11-20",
             "gpt-4o-2024-08-06",
@@ -105,7 +109,7 @@ class SubmitInputArgs(TypedDict, total=False):
         openai.NotGiven,
     ]
     presence_penalty: Union[float, None, openai.NotGiven]
-    reasoning_effort: Union[Literal["low", "medium", "high"], openai.NotGiven]
+    reasoning_effort: Union[Literal["low", "medium", "high"], None, openai.NotGiven]
     response_format: Union[
         openai.types.shared_params.response_format_text.ResponseFormatText,
         openai.types.shared_params.response_format_json_object.ResponseFormatJSONObject,