+++ docs/index.py
@@ -14,6 +14,11 @@ pagetitle: "chatlas"
 {readme_src}
 """
 
+# On Github, asset links are relative to the root directory,
+# but on the Quarto site, they are relative to the docs directory.
+# So, we need to adjust the asset links in the README.
+index_src = index_src.replace("docs/images/", "images/")
+
 index = docs_dir / "index.qmd"
 
 with open(index, "w") as f: