+++ tests/test_provider_anthropic.py
@@ -57,7 +57,13 @@ def test_anthropic_tool_variations():
     retryassert(run_simpleassert, retries=5)
 
     def run_parallelassert():
-        assert_tools_parallel(chat_fun)
+        # For some reason, at the time of writing, Claude 3.7 doesn't
+        # respond with multiple tools at once for this test (but it does)
+        # answer the question correctly with sequential tools.
+        def chat_fun2(**kwargs):
+            return ChatAnthropic(model="claude-3-5-sonnet-latest", **kwargs)
+
+        assert_tools_parallel(chat_fun2)
 
     retryassert(run_parallelassert, retries=5)
 
@@ -82,5 +88,9 @@ def test_data_extraction():
 
 def test_anthropic_images():
     chat_fun = ChatAnthropic
-    assert_images_inline(chat_fun)
+
+    def run_inlineassert():
+        assert_images_inline(chat_fun)
+
+    retryassert(run_inlineassert, retries=3)
     assert_images_remote_error(chat_fun)